Product statistics
Create a program that reads the name and price of various products. The program should ask if the user will continue. At the end, show: a) What is the total amount spent on the purchase. b) how many products cost more than R $ 1000, c) What is the cheapest product name?.